feat(retry): add higher-order lettable version of retry
