+++ src/operators/index.ts
@@ -55,6 +55,7 @@ export { race } from './race';
 export { reduce } from './reduce';
 export { repeat } from './repeat';
 export { repeatWhen } from './repeatWhen';
+export { retry } from './retry';
 export { refCount } from './refCount';
 export { scan } from './scan';
 export { subscribeOn } from './subscribeOn';
